Reduce code duplication for serving IO methods.
The downside is that the IO-ness of the server method is now part of the
interface, but callers don't really care because they are already in a
MonadIO.
